requirements.txt
Django==1.5.1
Markdown==2.3.1
South==0.8.1
argparse==1.2.1
dj-database-url==0.2.1
django-filter==0.6
django-nose==1.1
django-rest-framework-docs==0.1.6
django-rest-swagger==0.1.1
djangorestframework==2.3.5
docutils==0.10
gunicorn==17.5
jsonpickle==0.4.0
nose==1.3.0
psycopg2==2.5.1
requests==1.2.3
stevedore==0.9.1
virtualenv==1.9.1
virtualenv-clone==0.2.4
virtualenvwrapper==4.0
wsgiref==0.1.2
